feat(config): change the main function signature to include a config object
Reworked how values are passed in to the main functions to allow for an easier way to specify
options on how the smoother runs. This was becoming a problem with multiple optional function
parameters and is better handled as a object with optional props.

BREAKING CHANGE: Main function signature change